+++ CHANGELOG.md
@@ -6,5 +6,6 @@
 - Support of `/`
 - Support of `/lists`
 - Support of '/lists/<id>/interest-categories'
+- Support of 'lists/<id>/interest-categories/<id>/interests'
 - Support of '/lists/<id>/members'
 - `.count` support